requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
argcomplete==3.1.6
attrs==23.1.0
azure-functions==1.17.0
azure-functions-durable==1.2.8
backoff==2.2.1
censys==2.2.9
certifi==2023.11.17
charset-normalizer==3.3.2
frozenlist==1.4.0
furl==2.1.3
idna==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
orderedmultidict==1.0.1
Pygments==2.17.2
python-dateutil==2.8.2
requests==2.31.0
rich==13.7.0
six==1.16.0
urllib3==2.1.0
yarl==1.9.4
